pyproject.toml
[project]
name = "victus-backend"
version = "2.0.0"
description = "Advanced AI Personal Assistant with OpenAI"
authors = [
    {name = "Gaurav Vijay Jadhav"}
]
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
langchain = "~0.2.1"
langchain-openai = "^0.1.0"
langchain-community = "~0.2.1"
langchain-text-splitters = "~0.2.1"
python-dotenv = "^1.0.1"
tavily-python = "^0.3.3"
faiss-cpu = "^1.8.0"
pypdf = "^4.2.0"
python-docx = "^1.1.2"
sqlalchemy = "^2.0.30"
msal = "^1.28.0"
requests = "^2.32.3"
pyperclip = "^1.8.2"
faster-whisper = "^1.0.3"
piper-tts = "^1.2.0"
# torch = "^2.3.0"  # Install via conda: conda install pytorch -c pytorch
dateparser = "^1.2.2"
pyautogui = "^0.9.54"
pygetwindow = {version = "^0.0.9", markers = "sys_platform == 'win32'"}
httpx = "^0.27.0"
pydantic-settings = "^2.3.4"
slowapi = "^0.1.9"
alembic = "^1.13.0"
prometheus-client = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.7"
google-auth = "^2.25.0"
google-auth-oauthlib = "^1.2.0"
pillow = "^12.0.0"
fastapi-mail = "^1.6.1"
bcrypt = "3.2.2"
resend = "^2.19.0"
psycopg2-binary = "^2.9.9"

pgvector = "^0.2.5"
langfuse = "^2.30.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-api = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-requests = "^0.45b0"
opentelemetry-instrumentation-sqlalchemy = "^0.45b0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.3.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
